init from main. refactor: changed from float64 to int64 in timer and config. changed timerStart duration from 5 seconds to config doorDuration

+++ elevator/elevator.go
@@ -21,7 +21,7 @@ const (
 
 type Config struct {
 	ClearRequestVariant ClearRequestVariant
-	DoorOpenDurationS   float64
+	DoorOpenDurationS   int64
 }
 
 type Elevator struct { // add local ip mby?
